requirements.txt
selenium
aiohttp
pyyaml
